organized component files into the components folder
