+++ src/routes/AddAccount/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import LinkAccountModule from '../../LinkAccountModule'
+import LinkAccountModule from '../../Components/LinkAccountModule'
 
 export const Route = createFileRoute('/AddAccount/')({
   component: LinkAccountModule
