+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { MainModule } from '../MainModule'
+import { MainModule } from '../Components/MainModule'
 
 export const Route = createFileRoute('/')({
   component: MainModule
