--- src/statusbar.tsx
@@ -1,9 +1,0 @@
-function Statusbar(){
-return (
-    <div className="bg-TradeShareblue col-span-2">
-Feeds
-    </div>
-)
-}
-
-export default Statusbar
